Customized Database Management System
DBMS is consider as a software use to manage the database by handling the queery in SQL language.                          
In this project we implement the application which works like the DBMS.                                           
Programmatically we provide the logic for all types of queeries that we can write to maintain and manipulate the database.  To Apply all these customized queeries we consider a standard database.